Makefile: Use a single version of dmd, from path
The usage of two versions of DMD seems to be to work around old compiler issues.
Nowadays, DMD 2.079.0 is old enough (released 2018-03-01)
that it can be set as the mainline.
Finally, instead of used fixed locations that are dependent on the developer's setup,
simply use the DMD that should be in the PATH.